Extract code into a separate widget
