+++ src/vs/sessions/contrib/changes/browser/media/changesSummaryWidget.css
@@ -1,0 +1,32 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+.changes-summary-widget {
+	display: inline-flex;
+	gap: 4px;
+
+	.changes-summary-lines-added,
+	.changes-summary-lines-removed {
+		position: relative;
+		display: inline-grid;
+		overflow: hidden;
+		justify-items: end;
+		min-width: 2ch;
+
+		div {
+			grid-area: 1 / 1;
+			display: block;
+			will-change: transform, opacity;
+		}
+	}
+
+	.changes-summary-lines-added {
+		color: var(--vscode-chat-linesAddedForeground);
+	}
+
+	.changes-summary-lines-removed {
+		color: var(--vscode-chat-linesRemovedForeground);
+	}
+}
